+++ packages/orq-rc/src/models/operations/createprompt.ts
@@ -5816,7 +5816,7 @@ export const CreatePromptResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-01-31T07:05:40.814Z",
+    "2025-01-31T13:31:12.935Z",
   ).transform(v => new Date(v)),
   type: CreatePromptPromptsType$inboundSchema,
   versions: z.array(z.lazy(() => CreatePromptVersions$inboundSchema)),
@@ -5864,7 +5864,7 @@ export const CreatePromptResponseBody$outboundSchema: z.ZodType<
   promptConfig: z.lazy(() => CreatePromptPromptsPromptConfig$outboundSchema),
   metadata: z.lazy(() => CreatePromptPromptsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-01-31T07:05:40.814Z"))
+  updated: z.date().default(() => new Date("2025-01-31T13:31:12.935Z"))
     .transform(v => v.toISOString()),
   type: CreatePromptPromptsType$outboundSchema,
   versions: z.array(z.lazy(() => CreatePromptVersions$outboundSchema)),