+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -3967,7 +3967,7 @@ export const GetAllPromptsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-01-31T07:05:40.814Z",
+    "2025-01-31T13:31:12.935Z",
   ).transform(v => new Date(v)),
   type: GetAllPromptsType$inboundSchema,
   versions: z.array(z.lazy(() => GetAllPromptsVersions$inboundSchema)),
@@ -4015,7 +4015,7 @@ export const GetAllPromptsData$outboundSchema: z.ZodType<
   promptConfig: z.lazy(() => GetAllPromptsPromptConfig$outboundSchema),
   metadata: z.lazy(() => GetAllPromptsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-01-31T07:05:40.814Z"))
+  updated: z.date().default(() => new Date("2025-01-31T13:31:12.935Z"))
     .transform(v => v.toISOString()),
   type: GetAllPromptsType$outboundSchema,
   versions: z.array(z.lazy(() => GetAllPromptsVersions$outboundSchema)),