+++ packages/orq-rc/src/models/operations/updateprompt.ts
@@ -5884,7 +5884,7 @@ export const UpdatePromptResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-01-31T07:05:40.814Z",
+    "2025-01-31T13:31:12.935Z",
   ).transform(v => new Date(v)),
   type: UpdatePromptPromptsType$inboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptVersions$inboundSchema)),
@@ -5932,7 +5932,7 @@ export const UpdatePromptResponseBody$outboundSchema: z.ZodType<
   promptConfig: z.lazy(() => UpdatePromptPromptsPromptConfig$outboundSchema),
   metadata: z.lazy(() => UpdatePromptPromptsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-01-31T07:05:40.814Z"))
+  updated: z.date().default(() => new Date("2025-01-31T13:31:12.935Z"))
     .transform(v => v.toISOString()),
   type: UpdatePromptPromptsType$outboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptVersions$outboundSchema)),