+++ packages/orq-rc/src/models/operations/updatepromptsnippet.ts
@@ -6380,7 +6380,7 @@ export const UpdatePromptSnippetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-01-31T07:05:40.814Z",
+    "2025-01-31T13:31:12.935Z",
   ).transform(v => new Date(v)),
   type: UpdatePromptSnippetPromptSnippetsType$inboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptSnippetVersions$inboundSchema)),
@@ -6434,7 +6434,7 @@ export const UpdatePromptSnippetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string(),
   updatedById: z.string(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-01-31T07:05:40.814Z"))
+  updated: z.date().default(() => new Date("2025-01-31T13:31:12.935Z"))
     .transform(v => v.toISOString()),
   type: UpdatePromptSnippetPromptSnippetsType$outboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptSnippetVersions$outboundSchema)),