+++ packages/orq-rc/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v3.1.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.26 - packages/orq-rc
+
+## 2025-01-31 13:32:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.480.1 (2.499.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.27 - packages/orq-rc